сохранение периода в базе данных и вывод в форму
сейчас  есть только сохранение при создании новой таски

+++ Assets/[1] SOURCE/UI/TimeLabel.cs
@@ -19,7 +19,7 @@ public class TimeLabel : MonoBehaviour
 
     public TextMeshProUGUI LabelText;
 
-    public float TimeValue;
+    public int TimeValue;
 
     public Clock Clock;
